Cache Response
At this point, on slow screens we will get stuff load a whole lot faster, on lie-fie network, we still deliver stuff instead of no content at all. on offline we still deliver some content rather than complete failure.
However, the images and avatars did not load and if we remove force update and load the page, we get new posts, but when we go offline, we still get the old data and not the latest we got from when we were online. This is because looking at our code, we cached the html once at in 'install' time. So for now we are stuck with those cache messages

To complete the rest of the step to full have the app function offline
- We need to come up with a system for unobstrusive app updates
- We want to get the user using the latest version as quickly as possible
- We should continually update the cache of posts. So that we will not leave the user with the content we are originally with
- We need to cache the photos
- We need to cache the avatars