UnObstrusive app updates
After this commit is easier to now have in resources updates like change in css(maybe due to change in theme) or js, the user has to wait until pages that uses the current version go away. This works but it is not perfect yet. We want then to get it as soon as we make the change. We will see that in the next commit.
But the good thing now is that we did not disrupt the lifecycle of the service worker ;)
The concept of how your browser downloads updates in the background and waits till you restarts the browser to see new updates is thesame concept of how our site works now with the service worker

From this point one we will try not to use force update anymore which is how our user will experience the app, and get our app to work offline well
We need to work with the service worker to get it to pick up changes like change to css. Before now we only get new service worker when the sw.js changes.
This will be good because we developers change code all the time and we want the user to have it as fast as possible

We have to simulate a change in the service worker, then the browser will treat this updated worker as a new version. Because it is new, it will get its own install event where it will fetch the JS, HTML, updated css and put them in a new cache. We give this new cache a new name so that we will not interrupt the existing cache used by the old service worker and the page it controls. Once the old service worker is released(due to a page refresh), we delete the old cache so that the next page loads get the resources from the new cache; meaning it gets the latest css

Note: we left our css and images that way because we are in dev. look at the commets for best practise and why