Triggering an Update
To test our update trigger button, we basically changes the theme of our app, change the cache version name. then reload the page. Note that chaing cache name is not a good practise but the name of the resource instead

FINALLY, We have the user using the latest version as quickly as possible