Using IDB Cache
We set up database for witter and add posts to it
We created an object store 'wittr' as well as an index that sorts the messages by date.
We populated the database with messages from the webSockets

NOTE: During this commit, please mark the 'upload on reload' in the dev tools to see your changes right away for fatser development

WHAT WE WANT TO ACCOMPLISH
Using the IDB Cache and Display Entries
We want to create a database for witter that stores posts. At this point, when witrr loads via a service worker, it does so without going to the network. it fetches the page skeleton and assets straight from the cache.
We still have to go to the network for posts. We will change that. We will get post from the idb then display them before we go to the network. Then we connect the websocket to get updated posts.
By default, websocket by passes the http cache and database. So as new data comes in, we still display them, but we add them to our idb as well.