Finish implementing continually update the cache of posts
At a point, you will have thousands and thousands of items in a storage and the browser will be like they have had enough of you doing that and you are not allowed anymore storage(our websocket keep dumping data in there without proper clean up). So we need to make sure that we are only keeping what we need.

We want to make sure that our wittr storage has only 30 items in at at any given time.

To confirm what we did, we enabled force updates, refreshed the page and checked the storage to see that the data in there is no more than 30. We are good with that
The since we like this update, we now changed bumped the static cache of the servieWorker, disable forceUpdates to push the update as usual to our users

At this point we can launch our app. But we have to fix the image problem ;)