Cache the photos
We want to serve photos from the sw cache if they are there. otherwise we get them from the network and put them into the cache for the next time.

We use the storageUrl when matching and putting images into the image cache so that we only end up with one photo in the cache no mater how many different sizes of that same image are requested

We confirmed things are working fine, when we go to the 'Cache' section under the Application tab in dev tools, and we see the URLS are messing the width and the .jpg extension at the end. We took our app offline and still get images! Even  though we change the width of the browser and refresh the page, the images still load!

NOTE: at the monent, we just cached the images for posts and not the avatars for the users

When we get an item from the idb database we have to get the whole thing out in one lump, then convert it to image data then add it to the page.
In contrast, if we get the image from our SW cache, it will stream the data. So we dont need to wait for the whole thing before we display anything. This is more memory efficient and leads to faster renders even if the data is coming from the disk. So we will use the Cache API to cache images rather than indexedDB. NOTE that we put the image in a cache that is different from where we store our static content like page skeletons, css, etc. if the user makes a request for an image that we already have cached, we simply return it.