Caching Avatars
Caching avatars is similar to caching phots but the difference is that some people like to change their photos a lot and we dont want them to be stuck with the old version. So when we fetch a particular avatar from the cache, we will also go to the network to update the cache again for the next fetch

After this commit, we can still see user avatars even when the app is offline!

Our app is not full functional in offline
Congratulations!!